+++ pkg/vpcmodel/externalNetwork.go
@@ -3,8 +3,6 @@ package vpcmodel
 import (
 	"errors"
 	"fmt"
-	"strings"
-
 	"github.com/np-guard/vpc-network-config-analyzer/pkg/common"
 )
 
@@ -155,23 +153,4 @@ func isEntirePublicInternetRange(nodes []Node) (bool, error) {
 		return false, err
 	}
 	return nodesRanges.Equal(allInternetRagnes), nil
-}
-
-func (g *groupedExternalNodes) String() string {
-	// 1. Created a list of IPBlocks
-	cidrList := make([]string, len(*g))
-	for i, n := range *g {
-		cidrList[i] = n.Cidr()
-	}
-	ipbList, _, err := ipStringsToIPblocks(cidrList)
-	if err != nil {
-		return ""
-	}
-	// 2. Union all IPBlocks in a single one; its intervals will be the cidr blocks or ranges that should be printed, after all possible merges
-	unionBlock := &common.IPBlock{}
-	for _, ipBlock := range ipbList {
-		unionBlock = unionBlock.Union(ipBlock)
-	}
-	// 3. print a list s.t. each element contains either a single cidr or an ip range
-	return strings.Join(unionBlock.ListToPrint(), commaSepartor)
 }
